refactor: 💡 remove route-action usage for scopes in desktop
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-12621
