+++ ui/desktop/app/controllers/error.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ErrorController extends Controller {
+  @controller('application') appUtilities;
+}
