+++ ui/desktop/app/controllers/scopes/scope/authenticate/method/oidc.js
@@ -14,7 +14,7 @@ export default class ScopesScopeAuthenticateMethodOidcController extends Control
   // =attributes
 
   /**
-   * Authetnication URL for the pending OIDC flow, if any.
+   * Authentication URL for the pending OIDC flow, if any.
    * @type {?string}
    */
   get authURL() {
