+++ ui/desktop/app/controllers/scopes/scope/projects/error.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeProjectsErrorController extends Controller {
+  @controller('application') appUtilities;
+}
