+++ ui/desktop/app/templates/application.hbs
@@ -26,21 +26,21 @@
             @style='warning'
             class='button-window-close'
             title={{t 'actions.close'}}
-            {{on 'click' (route-action 'close')}}
+            {{on 'click' this.close}}
           />
           <Rose::Button
             @iconOnly='flight-icons/svg/square-16'
             @style='secondary'
             class='button-window-fullscreen'
             title={{t 'actions.fullscreen'}}
-            {{on 'click' (route-action 'toggleFullScreen')}}
+            {{on 'click' this.toggleFullScreen}}
           />
           <Rose::Button
             @iconOnly='flight-icons/svg/minus-16'
             @style='secondary'
             class='button-window-minimize'
             title={{t 'actions.minimize'}}
-            {{on 'click' (route-action 'minimize')}}
+            {{on 'click' this.minimize}}
           />
         {{/if}}
 
@@ -50,7 +50,7 @@
             @icon='flight-icons/svg/user-circle-16'
             as |dropdown|
           >
-            <dropdown.button {{on 'click' (route-action 'invalidateSession')}}>
+            <dropdown.button {{on 'click' this.invalidateSession}}>
               {{t 'actions.deauthenticate'}}
             </dropdown.button>
