+++ ui/desktop/app/templates/cluster-url.hbs
@@ -10,7 +10,7 @@
   >
     <Rose::Form
       class='full-width'
-      @onSubmit={{route-action 'setClusterUrl' this.clusterUrl}}
+      @onSubmit={{fn this.setClusterUrl this.clusterUrl}}
       @cancel={{false}}
       @disabled={{(is-loading)}}
       as |form|
