+++ ui/desktop/tests/unit/controllers/scopes/scope/authenticate/method/index-test.js
@@ -1,0 +1,17 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'ember-qunit';
+
+module(
+  'Unit | Controller | scopes/scope/authenticate/method/index',
+  function (hooks) {
+    setupTest(hooks);
+
+    // TODO: Replace this with your real tests.
+    test('it exists', function (assert) {
+      let controller = this.owner.lookup(
+        'controller:scopes/scope/authenticate/method/index',
+      );
+      assert.ok(controller);
+    });
+  },
+);
